// src/models/models.ts
export interface ITag {
    name: string;
    color: string;
}